rename:  Notification을 FCM 로 변경 #39
- 불필요한 알람 설정 페이지 삭제
